型専用ファイルをindexにまとめる
